--- src/test/java/com/trycloud/tests/UserStory1/TestCase2.java
@@ -1,6 +1,0 @@
-package com.trycloud.tests.UserStory1;
-
-import com.trycloud.utilities.Driver;
-
-public class TestCase2 {
-}
