+++ src/test/java/com/trycloud/tests/user_story_1/test1.java
@@ -1,4 +1,4 @@
-package com.trycloud.tests.UserStory1;
+package com.trycloud.tests.user_story_1;
 
 import com.trycloud.utilities.BrowserUtils;
 import com.trycloud.utilities.WebDriverFactory;
@@ -7,11 +7,10 @@ import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.testng.Assert;
 import org.testng.annotations.Test;
-import org.testng.asserts.Assertion;
 
 import java.util.concurrent.TimeUnit;
 
-public class test2 {
+public class test1 {
 
     WebDriver driver;
 
@@ -22,7 +21,7 @@ public class test2 {
         driver= WebDriverFactory.getDriver("chrome");
 
         //2- go to  http://qa.trycloud.net/index.php/login?clear=1
-        driver.get(" http://qa3.trycloud.net/index.php/login");
+        driver.get("http://qa3.trycloud.net/");
 
         //3-maximize the window
         driver.manage().window().maximize();
@@ -36,42 +35,39 @@ public class test2 {
         userName.sendKeys("User21");
 
         BrowserUtils.sleep(1);
+        //6- Enter valid password
 
         WebElement password = driver.findElement(By.xpath("//input[@id='password']"));
         password.click();
-        password.sendKeys("Userpass33333");
+        password.sendKeys("Userpass123");
 
         BrowserUtils.sleep(1);
+        // 7- Click login button
 
         WebElement loginButton= driver.findElement(By.xpath("//input[@id='submit-form']"));
         loginButton.click();
 
         BrowserUtils.sleep(1);
+        //8- Verify the URL is changed to homepage’s url
+
 
         String actualUrl= driver.getCurrentUrl();
-        String expectedUrl= "http://qa3.trycloud.net/index.php/login";
+        String expectedUrl= "http://qa3.trycloud.net/index.php/apps/dashboard/";
 
 
-    /*    if (actualUrl.equals(expectedUrl)){
-            System.out.println("Url verifed");
-        }else{
-            System.out.println("Url verified Failed");
-        }
-*/
-
+        Assert.assertTrue(actualUrl.equals(expectedUrl));
+        driver.close();
 
 
 
     }
 }
-/*Test case #2 - verify user cannot login with invalid credentials
 
-        1. Navigate to login page http://qa.trycloud.net/index.php/login?clear=1
+/*Test case #1 =====> verify user login successfully
 
-        2. Enter valid username invalid password
-
-        3. Click login button
-
-        4. Message “Wrong username or password.” should be displayed
-
-        5. Page title and url should be same  */
+1. Navigate to login page http://qa.trycloud.net/index.php/login?clear=1
+2. Enter valid username
+3. Enter valid password
+4. Click login button
+5. Verify the URL is chawed to homepage’s url
+(Put this test case in testBase class with an annotation so it runs for all the test case)   */
