+++ src/test/java/com/trycloud/tests/user_story_1/test2.java
@@ -1,4 +1,4 @@
-package com.trycloud.tests.UserStory1;
+package com.trycloud.tests.user_story_1;
 
 import com.trycloud.utilities.BrowserUtils;
 import com.trycloud.utilities.WebDriverFactory;
@@ -7,11 +7,10 @@ import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.testng.Assert;
 import org.testng.annotations.Test;
-import org.testng.asserts.Assertion;
 
 import java.util.concurrent.TimeUnit;
 
-public class test1 {
+public class test2 {
 
     WebDriver driver;
 
@@ -21,8 +20,8 @@ public class test1 {
         //open the crhrome browser
         driver= WebDriverFactory.getDriver("chrome");
 
-        //2- go to  http://qa.trycloud.net/index.php/login?clear=1
-        driver.get("http://qa2.trycloud.net/");
+        //2- go to  http://qa3.trycloud.net/index.php/login
+        driver.get(" http://qa3.trycloud.net/index.php/login");
 
         //3-maximize the window
         driver.manage().window().maximize();
@@ -36,40 +35,47 @@ public class test1 {
         userName.sendKeys("User21");
 
         BrowserUtils.sleep(1);
+        // 6. Enter valid username invalid password
+
 
         WebElement password = driver.findElement(By.xpath("//input[@id='password']"));
         password.click();
-        password.sendKeys("Userpass123");
+        password.sendKeys("Userpass33333");
 
         BrowserUtils.sleep(1);
+        //  7. Click login button
 
         WebElement loginButton= driver.findElement(By.xpath("//input[@id='submit-form']"));
         loginButton.click();
 
         BrowserUtils.sleep(1);
+        //8. Message “Wrong username or password.” should be displayed
+
+        WebElement wrongPassWarning= driver.findElement(By.xpath("//p[@class='warning wrongPasswordMsg']"));
+        Assert.assertTrue(wrongPassWarning.isDisplayed());
+        //9. Page title and url should be same
+
 
         String actualUrl= driver.getCurrentUrl();
-        String expectedUrl= "http://qa2.trycloud.net/index.php/login";
+        String expectedUrl= "http://qa3.trycloud.net/index.php/login?user=User21";
+        Assert.assertTrue(actualUrl.equals(expectedUrl));
+        driver.close();
 
 
-    /*    if (actualUrl.equals(expectedUrl)){
-            System.out.println("Url verifed");
-        }else{
-            System.out.println("Url verified Failed");
-        }
-*/
 
 
 
 
     }
 }
+/*Test case #2 - verify user cannot login with invalid credentials
 
-/*Test case #1 =====> verify user login successfully
+        1. Navigate to login page http://qa2.trycloud.net
 
-1. Navigate to login page http://qa.trycloud.net/index.php/login?clear=1
-2. Enter valid username
-3. Enter valid password
-4. Click login button
-5. Verify the URL is chawed to homepage’s url
-(Put this test case in testBase class with an annotation so it runs for all the test case)   */
+        2. Enter valid username invalid password
+
+        3. Click login button
+
+        4. Message “Wrong username or password.” should be displayed
+
+        5. Page title and url should be same  */
