+++ src/test/java/com/trycloud/tests/user_story_3/Test_case_4.java
@@ -1,11 +1,9 @@
 package com.trycloud.tests.user_story_3;
 
-import com.trycloud.tests.UserStory1.LoginToWebsite;
-import com.trycloud.utilities.ConfigurationReader;
+import com.trycloud.tests.user_story_1.LoginToWebsite;
 import com.trycloud.utilities.Driver;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebElement;
-import org.testng.Assert;
 import org.testng.annotations.Test;
 
 public class Test_case_4 {
