+++ src/test/java/com/trycloud/tests/user_story_3/Test_case_6.java
@@ -1,17 +1,11 @@
 package com.trycloud.tests.user_story_3;
 
-import com.google.common.base.Verify;
-import com.trycloud.tests.UserStory1.LoginToWebsite;
-import com.trycloud.utilities.ConfigurationReader;
+import com.trycloud.tests.user_story_1.LoginToWebsite;
 import com.trycloud.utilities.Driver;
-import io.github.bonigarcia.wdm.WebDriverManager;
 import org.openqa.selenium.By;
 import org.openqa.selenium.Keys;
-import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
-import org.openqa.selenium.chrome.ChromeDriver;
 import org.testng.Assert;
-import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 
 import java.util.concurrent.TimeUnit;
