+++ src/test/java/com/trycloud/tests/user_story_3/Test_case_9.java
@@ -1,6 +1,6 @@
 package com.trycloud.tests.user_story_3;
 
-import com.trycloud.tests.UserStory1.LoginToWebsite;
+import com.trycloud.tests.user_story_1.LoginToWebsite;
 import com.trycloud.utilities.BrowserUtils;
 import com.trycloud.utilities.Driver;
 import org.openqa.selenium.By;
