+++ src/test/java/com/trycloud/tests/user_story_5/TestCase5_2.java
@@ -1,14 +1,11 @@
-package com.trycloud.tests.TestCase5;
+package com.trycloud.tests.user_story_5;
 
 import com.github.javafaker.Faker;
-import com.trycloud.tests.UserStory1.LoginToWebsite;
+import com.trycloud.tests.user_story_1.LoginToWebsite;
 import com.trycloud.utilities.BrowserUtils;
 import com.trycloud.utilities.Driver;
-import io.github.bonigarcia.wdm.WebDriverManager;
 import org.openqa.selenium.By;
-import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
-import org.openqa.selenium.chrome.ChromeDriver;
 import org.testng.Assert;
 import org.testng.annotations.Test;
