+++ src/test/java/com/trycloud/tests/user_story_7/testCase7.java
@@ -1,6 +1,6 @@
-package com.trycloud.tests.Userstory7;
+package com.trycloud.tests.user_story_7;
 
-import com.trycloud.tests.UserStory1.LoginToWebsite;
+import com.trycloud.tests.user_story_1.LoginToWebsite;
 import com.trycloud.utilities.Driver;
 import org.openqa.selenium.By;
 import org.openqa.selenium.Keys;
